Refactor CausalInference again
