+++ requirements-dev.txt
@@ -6,5 +6,4 @@ pyparsing
 mock
 nose
 wrapt
-pytorch
 statsmodels
