+++ requirements.txt
@@ -4,5 +4,4 @@ scipy
 pandas
 pyparsing
 wrapt
-pytorch
 statsmodels
